refactor: moved detail page into common widget
