+++ lib/view/common_widgets/detail_animal_page.dart
