+++ lib/view/common_widgets/detail_plant_page.dart
